webapp/requirements.txt
flask==3.1.0
jsonify==0.5
mysql-connector-python==9.1.0
